pyproject.toml
[project]
name = "gemini-live"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "eel>=0.18.2",
    "google-genai>=1.41.0",
    "mss>=10.1.0",
    "opencv-python>=4.12.0.88",
    "pillow>=11.3.0",
    "pyaudio>=0.2.14",
]
